Clarificacion de codigo, y arreglo de distintas cosas

+++ src/main/java/com/uade/grupo4/backend_ecommerce/service/implementations/CartService.java
@@ -8,7 +8,6 @@ import com.uade.grupo4.backend_ecommerce.controller.dto.UserDto;
 import com.uade.grupo4.backend_ecommerce.repository.CartItemRepository;
 import com.uade.grupo4.backend_ecommerce.repository.CartRepository;
 import com.uade.grupo4.backend_ecommerce.repository.ProductRepository;
-import com.uade.grupo4.backend_ecommerce.repository.mapper.CartMapper;
 import com.uade.grupo4.backend_ecommerce.repository.entity.Cart;
 import com.uade.grupo4.backend_ecommerce.repository.entity.CartItem;
 import com.uade.grupo4.backend_ecommerce.repository.entity.Product;
@@ -35,8 +34,7 @@ public class CartService implements CartServiceInterface {
     public CartDto addProductToCart(Long carritoID, Long productId, int quantity) {
         Cart cart = cartRepository.findById(carritoID).orElseThrow();
         Product product = productRepository.findById(productId).orElseThrow();
-
-        CartItem existingItem = cartItemRepository.findByCartAndProduct(cart.getId(), productId);
+        CartItem existingItem = cartItemRepository.findByCartIdAndProductId(cart.getId(), productId);
         if (existingItem != null) {
             if (product.getQuantity() <(existingItem.getQuantity() + quantity)){
                 throw  new RuntimeException("No hay suficiente stock de "+product.getTitle()+"para agregar"+quantity+
@@ -52,16 +50,16 @@ public class CartService implements CartServiceInterface {
             CartItem newItem = new CartItem(carritoID, cart, product, quantity);
             cartItemRepository.save(newItem);
         }
-        cart.setTotal(cart.getTotal() + 10 * quantity);
+        cart.setTotal(cart.getTotal() + product.getPrice() * quantity);
         cartRepository.save(cart);
         return new CartDto(carritoID,new UserDto(1,"Federico","fed","fe","fe"), (List<CartItemDto>) existingItem,"10");
-    }
+    }//poner el mapper toDTO
 
     public CartDto removeProductFromCart(Long carritoID, Long productId, int quantity) throws Exception {
         Cart cart = cartRepository.findById(carritoID).orElseThrow();
         Product product = productRepository.findById(productId).orElseThrow();
 
-        CartItem existingItem = cartItemRepository.findByCartAndProduct(cart.getId(), productId);
+        CartItem existingItem = cartItemRepository.findByCartIdAndProductId(cart.getId(), productId);
         if (existingItem != null) {
             int newQuantity = existingItem.getQuantity() - quantity;
             if (newQuantity < 0) {
@@ -76,32 +74,36 @@ public class CartService implements CartServiceInterface {
         else{
             throw new RuntimeException("El producto"+product.getTitle()+"no existe en el carrito de usted");
         }//preguntar si hay que crear excepciones
-        cart.setTotal(cart.getTotal() - 10 * quantity);
+        cart.setTotal(cart.getTotal() - product.getPrice() * quantity);
         cartRepository.save(cart);
         return new CartDto(carritoID,new UserDto(1,"Federico","fed","fe","fe"), (List<CartItemDto>) existingItem,"10");
-    }
+    }//poner el mapper toDTO
 
 
-    public void emptyCart(Long cartId) {
+    public boolean emptyCart(Long cartId) {
         Cart cart = cartRepository.findById(cartId).orElseThrow();
-        /*List<CartItem> cartItems = cartItemRepository.findByCart(cartId);
+        List<CartItem> cartItems = cartItemRepository.findByCartId(cartId);
+        if(cartItems.isEmpty()){
+            return false;
+        }
         for (CartItem item : cartItems) {
-            //Long productId = item.getProduct().getId();
+            Long productId = item.getProduct().getId();
             int quantity = item.getQuantity();
             //Product product = productRepository.findById(productId).orElseThrow();
             //product.setQuantity(product.getQuantity() + quantity);
-            //productRepository.save(product); // QUE LO PASE A REPOSITORIO
+            //productRepository.save(product);
         }
         cartItemRepository.deleteAll(cartItems);
-        //cart.setItems(new ArrayList<>());
-        cartRepository.save(cart);*/
+        cart.setItems(new ArrayList<>());
+        cartRepository.save(cart);
+        return true;
     }
 
     public float checkoutCart(Long cartId) {
         Cart cart = cartRepository.findById(cartId).orElseThrow();
-       /* List<CartItem> cartItems = cartItemRepository.findByCart(cartId);
+        List<CartItem> cartItems = cartItemRepository.findByCartId(cartId);
         for (CartItem item : cartItems) {
-            //Long productId = item.getProduct().getId();
+            Long productId = item.getProduct().getId();
             int quantity = item.getQuantity();
             //Product product = productRepository.findById(productId).orElseThrow();
            // if (product.getQuantity() < quantity) {
@@ -110,9 +112,9 @@ public class CartService implements CartServiceInterface {
             //product.setQuantity(product.getQuantity() - quantity);
             //productRepository.save(product);
             cart.setTotal(cart.getTotal() + 10 * quantity);
-        }*/
-        //cartItemRepository.deleteAll(cartItems);
-        //cart.setItems(new ArrayList<>());
+        }
+        cartItemRepository.deleteAll(cartItems);
+        cart.setItems(new ArrayList<>());
         cartRepository.save(cart);
         return cart.getTotal();
